fix(scoring): allow data to be copied directly from twitter (#131)
I've updated the config files for each year to require only two files:

1. `picks.hockey.js` - holds our picks for each round for each competitor
2. `results.hockey.js` - holds the results for each round

See the diff for examples of each file.

The changes have been made for two reasons:

1. Simplify data entry. It was a pay to add all of the data separately for each round and everyone's picks
2. Reduce transcription errors. Mapping a competitor's pick to the game name to the correct team name with the correct games for the correct was massively error prone. I...uhh...may have made some mistakes in the past. This changeset should almost entirely eliminate the transcription errors because it is optimized for copy/pasting a competitor's picks

A competitor's picks looks something like:

```
@thoronas: COL 5, VGK 7, TOR 6, EDM 7, PIT 7, BOS 6, CAR 5, FLA 7
```

The parser for the pick line is meant to be forgiving. It's looking for 2-3 letters followed by a number.

The biggest challenge we'll have initially is building out the `team-codes.json` file. Everyone kinda uses their own abbreviations. We need to map them as they come up. The whole site will error when it happens, so we'll recognize and fix it quickly.

I've tested the new changes against the old version of the site and it produces the same results. Tests are still needed though.